// src/app/api/generated/model/focusSessionResDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FocusSessionResDto { 
    id?: number;
    title?: string;
    timer?: string;
    userId?: number;
}